Cannot make all files together with ProgramDisplay files. However cmd line parsing works.
