+++ CAN Wrapper/ProgramDisplay.c
@@ -1,0 +1,15 @@
+
+#include <string.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+//Class includes..
+#include "ProgramDisplay.h"
+
+void StartupInfo()
+{
+        printf("\n\t -HOW TO START READER-\n");
+        printf("\t\t Reader [-r] (RAW CAN DATA Wrapper mode\n");
+        printf("\t\t        [-j] (J1939 CAN DATA Wrapper mode\n");
+        printf("\t\t        [-s] (Needed actions to have both modes working\n");
+}
