+++ CAN Wrapper/ProgramDisplay.h
@@ -1,0 +1,11 @@
+/*
+* FILE        : porgramDisplay.h
+* PROJECT     : CAN Reader
+* DATE        : 2020-02-21
+* DESCRIPTION : Contains definitions needed for the programDisplay class. 
+*/
+
+#include <stdio.h>
+#include <stdlib.h>
+
+void StartupInfo();
